Create forms.html
Name - makes it easy to refer to(especially while coding)
Type - what we want user to input(type of input)
Placeholder - The placeholder attribute specifies a short hint that describes the expected value of an input field (e.g. a sample value or a short description of the expected format).